refactor: Rule 인터페이스 생성
