+++ src/main/kotlin/omok/model/Game.kt
@@ -1,6 +1,7 @@
 package omok.model
 
 import omok.model.position.Position
+import omok.model.rule.Rule
 
 class Game(val board: Board, private val rule: Rule) {
     var lastStone: Stone? = null
