+++ src/test/kotlin/omok/model/GameTest.kt
@@ -3,13 +3,14 @@ package omok.model
 import omok.model.position.Col
 import omok.model.position.Position
 import omok.model.position.Row
+import omok.model.rule.RenjuRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 class GameTest {
     @Test
     fun `처음에는 흑이 수를 둔다`() {
-        val game = Game(Board(), Rule())
+        val game = Game(Board(), RenjuRule())
 
         val actual: Color = game.chooseTurn()
         val expected: Color = Color.BLACK
@@ -19,7 +20,7 @@ class GameTest {
 
     @Test
     fun `흑이 수를 둔 후에는 백이 수를 둔다`() {
-        val game = Game(Board(), Rule()).apply { processTurn(Position(Col(1), Row(1)), Color.BLACK) }
+        val game = Game(Board(), RenjuRule()).apply { processTurn(Position(Col(1), Row(1)), Color.BLACK) }
 
         val actual: Color = game.chooseTurn()
         val expected: Color = Color.WHITE
@@ -29,7 +30,7 @@ class GameTest {
 
     @Test
     fun `백이 수를 둔 후에는 흑이 수를 둔다`() {
-        val game = Game(Board(), Rule()).apply { processTurn(Position(Col(1), Row(1)), Color.WHITE) }
+        val game = Game(Board(), RenjuRule()).apply { processTurn(Position(Col(1), Row(1)), Color.WHITE) }
 
         val actual: Color = game.chooseTurn()
         val expected: Color = Color.BLACK
@@ -40,7 +41,7 @@ class GameTest {
     @Test
     fun `오목이 완성되면 게임이 종료된다 1`() {
         val game =
-            Game(Board(), Rule()).apply {
+            Game(Board(), RenjuRule()).apply {
                 processTurn(Position(Col(1), Row(1)), Color.WHITE)
                 processTurn(Position(Col(1), Row(2)), Color.WHITE)
                 processTurn(Position(Col(1), Row(3)), Color.WHITE)
@@ -59,7 +60,7 @@ class GameTest {
     @Test
     fun `오목이 완성되면 게임이 종료된다 2`() {
         val game =
-            Game(Board(), Rule()).apply {
+            Game(Board(), RenjuRule()).apply {
                 processTurn(Position(Col(13), Row(5)), Color.BLACK)
                 processTurn(Position(Col(12), Row(6)), Color.BLACK)
                 processTurn(Position(Col(11), Row(7)), Color.BLACK)
@@ -78,7 +79,7 @@ class GameTest {
     @Test
     fun `오목이 완성되지 않았으면 게임이 진행 중이다`() {
         val game =
-            Game(Board(), Rule()).apply {
+            Game(Board(), RenjuRule()).apply {
                 processTurn(Position(Col(8), Row(8)), Color.WHITE)
                 processTurn(Position(Col(8), Row(9)), Color.WHITE)
                 processTurn(Position(Col(9), Row(8)), Color.WHITE)
